Numerous sound fixes and ending2 port start
